[SceneTime] Scene and Time refactoring
Scene cleanup and refactoring
Delta time changes

+++ Source/Engine.h
@@ -25,7 +25,7 @@ namespace QwerkE {
 
 		void NewFrame();
 		void PollInput();
-		void Update(double deltatime);
+		void Update(float deltatime);
 		void Draw();
 	}
